pyproject.toml
[project]
name = "dreamscape"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.129.0",
    "gradio>=6.5.1",
    "httpx>=0.28.1",
    "langgraph>=1.0.8",
    "litellm>=1.81.13",
    "loguru>=0.7.3",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.13.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.1",
    "sqlalchemy[asyncio]>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.1",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "N",   # pep8-naming
    "UP",  # pyupgrade (modern Python syntax)
    "B",   # flake8-bugbear (catches common bugs)
]
ignore = [
    "B008",  # FastAPI uses function calls (Depends) in defaults for dependency injection
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
